app/backend/requirements.txt
fastapi
uvicorn
python-chess
pydantic